requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0

# Existing dependencies for RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
neo4j==5.15.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
openai==1.6.1
python-dotenv==1.0.0
qdrant-client 1.6.4
